Added implementation for the Blow Up Unicorn.
This involved adding an enum to store the current location (Deck, Hand,
Stable, or Discard Pile) of the given card and appropriate method and test changes.
Also required adding a third value to some methods that only require it
for consistency: these are given the variable name "trash".

Will need to write more tests once the destroying of cards is properly
implemented